src/minimal-local/pyproject.toml
[project]
name = "ai-shield-backend"
version = "0.1.0"
description = "AI Shield Intelligence - Minimal Local Profile Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "minio>=7.2.3",
    "pyyaml>=6.0.1",
    "watchfiles>=0.21.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/models", "backend/api", "backend/services", "backend/collectors", "backend/scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["backend"]
omit = ["backend/tests/*", "backend/scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
